Convert Chevron component from .js to .tsx
